First-stage refactor of moving the loading/importing sequence to the new MVC structure.
