+++ web-client/public/js/grnstate.js
@@ -1,5 +1,18 @@
+let currentNetwork = null;
+
 export const grnState = {
     name: null,
-    network: null,
+
+    newNetwork: false,
+
+    get network() {
+        return currentNetwork;
+    },
+
+    set network(network) {
+        currentNetwork = network;
+        this.newNetwork = true;
+    },
+
     dashedLine: false,
 };
