+++ web-client/public/js/update-app.js
@@ -1,5 +1,6 @@
 import { drawGraph } from "./graph";
 import { uploadState } from "./upload";
+import { displayWarnings } from "./warnings";
 
 import {
   GREY_EDGES_DASHED_MENU,
@@ -14,7 +15,35 @@ const refreshApp = () => {
     }
 };
 
+var displayNetwork = function (network, name) {
+    uploadState.nodeColoring.reload(network, name);
+    if (document.getElementById("zoomSlider").disabled) {
+        document.getElementById("zoomSlider").disabled = false;
+    }
+
+    uploadState.currentNetwork = network;
+    console.log("Network: ", network); // Display the network in the console
+    $("#graph-metadata").html(network.genes.length + " nodes<br>" + network.links.length + " edges");
+
+    if (network.warnings.length > 0) {
+        displayWarnings(network.warnings);
+    }
+
+    $("#fileName").text(name); // Set the name of the file to display in the top bar
+    $("input[type='range']").off("input"); // I have no idea why I do this. Investigate later.
+
+    // If more things need to be turned off, we'll add them to this array
+    [ "#resetSliders", "#resetSlidersMenu", "#undoReset", "#undoResetMenu" ].forEach(function (selector) {
+        $(selector).off("click");
+    });
+};
+
 export const updateApp = grnState => {
+    if (grnState.newNetwork) {
+        displayNetwork(grnState.network, grnState.name);
+        refreshApp();
+    }
+
     if (grnState.dashedLine) {
         $(GREY_EDGES_DASHED_MENU + " span").addClass("glyphicon-ok");
         $(GREY_EDGES_DASHED_MENU).prop("checked", "checked");
